refactor: split geometry helper types
