+++ src/CADShared/Cad/Geometry/SpatialIndex/QuadTree/QuadTreeFindMode.cs
@@ -1,0 +1,24 @@
+namespace Fs.Fox.Cad;
+
+/// <summary>
+/// 四叉树查找方向
+/// </summary>
+public enum QuadTreeFindMode
+{
+    /// <summary>
+    /// 上
+    /// </summary>
+    Top    = 1,
+    /// <summary>
+    /// 下
+    /// </summary>
+    Bottom = 2,
+    /// <summary>
+    /// 左
+    /// </summary>
+    Left   = 4,
+    /// <summary>
+    /// 右
+    /// </summary>
+    Right  = 8,
+}
