+++ src/CADShared/Cad/Geometry/SpatialIndex/QuadTree/QuadTreeSelectMode.cs
@@ -14,26 +14,3 @@ public enum QuadTreeSelectMode
     /// </summary>
     Contains,     
 }
-
-/// <summary>
-/// 四叉树查找方向
-/// </summary>
-public enum QuadTreeFindMode
-{
-    /// <summary>
-    /// 上
-    /// </summary>
-    Top    = 1,  
-    /// <summary>
-    /// 下
-    /// </summary>
-    Bottom = 2,  
-    /// <summary>
-    /// 左
-    /// </summary>
-    Left   = 4, 
-    /// <summary>
-    /// 右
-    /// </summary>
-    Right  = 8,  
-}
